.devcontainer/apps/ai-travel-advisor/requirements.txt
fastapi == 0.115.12
uvicorn == 0.34.2
traceloop-sdk == 0.36.1
ollama~=0.4.8
langchain==0.3.23
langchain-weaviate==0.0.4
langchain-community==0.3.27
langchain-ollama==0.3.2
beautifulsoup4==4.13.4
lxml==5.3.2